20221205 参照する外部ファイルをgit管理するように変更

+++ inc/Release/DirectXTK12/DescriptorHeap.h
@@ -13,9 +13,6 @@
 #include <d3d12_xs.h>
 #elif (defined(_XBOX_ONE) && defined(_TITLE)) || defined(_GAMING_XBOX)
 #include <d3d12_x.h>
-#elif defined(USING_DIRECTX_HEADERS)
-#include <directx/d3d12.h>
-#include <dxguids/dxguids.h>
 #else
 #include <d3d12.h>
 #endif
@@ -49,9 +46,7 @@ namespace DirectX
             size_t count) noexcept(false) :
             DescriptorHeap(device,
                 D3D12_DESCRIPTOR_HEAP_TYPE_CBV_SRV_UAV,
-                D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE, count)
-        {
-        }
+                D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE, count) {}
 
         DescriptorHeap(DescriptorHeap&&) = default;
         DescriptorHeap& operator=(DescriptorHeap&&) = default;
@@ -123,7 +118,7 @@ namespace DirectX
         size_t Count() const noexcept { return m_desc.NumDescriptors; }
         unsigned int Flags() const noexcept { return m_desc.Flags; }
         D3D12_DESCRIPTOR_HEAP_TYPE Type() const noexcept { return m_desc.Type; }
-        uint32_t Increment() const noexcept { return m_increment; }
+        size_t Increment() const noexcept { return m_increment; }
         ID3D12DescriptorHeap* Heap() const noexcept { return m_pHeap.Get(); }
 
         static void __cdecl DefaultDesc(
@@ -195,9 +190,7 @@ namespace DirectX
             size_t reserve = 0) noexcept(false) :
             DescriptorPile(device,
                 D3D12_DESCRIPTOR_HEAP_TYPE_CBV_SRV_UAV,
-                D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE, count, reserve)
-        {
-        }
+                D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE, count, reserve) {}
 
         DescriptorPile(const DescriptorPile&) = delete;
         DescriptorPile& operator=(const DescriptorPile&) = delete;
